added util.py for all functions
